Obradjen format racuna 4, s tim da nisu u potpunosti zavrsene sve funkcije koje su potrebne za potpuno obradu, postoji mala komplikacija prilikom vracanja nizova koja ce u sto skorijem roku biti ispravljena i nije u potpunosti zavrsena funkcija za obradu mjeseca. Moguce je da taj nacin nije najbolje osmisljen pa ako neko ima bolji prijednog, neka predlozi. Kolege mogu pogledati koje funkcije postoje u ObradaRacuna.h i ObradaRacuna.c fajlovima, kao i u fajlovima ProvjeraFormata4.h i ProvjeraFormata4.c sto ih moze usmjeriti kako da obrade ostale formate!

+++ Strukture.h
@@ -14,8 +14,8 @@ typedef struct _korisnik
 typedef struct _proizvod {
 	char naziv[21];
 	int kolicina;
-	int cijena;
-	int ukupno;
+	double cijena;
+	double ukupno;
 }PROIZVOD;
 
 typedef struct _kupac {
@@ -26,7 +26,7 @@ typedef struct _kupac {
 typedef struct _mjesec {
 	int mjesec;
 	int godina;
-	KUPAC* ukupnaProdaja; //svi kupci koji su kupovali u odedjenom mjesecu 
+	KUPAC* ukupnaProdaja; //svi kupci koji su kupovali u odedjenom mjesecu
 }MJESEC;
 
 #endif // !STRUKTURE_H
